Update pull requests
