+++ account/admin.py
@@ -1,6 +1,3 @@
 from django.contrib import admin
-# from .models import Profile
-# Register your models here.
 
-# admin.site.register(Profile)
