+++ account/urls.py
@@ -4,6 +4,4 @@ from .views import  RequestPasswordResetEmail
 urlpatterns = [
     path('request-reset-email/', RequestPasswordResetEmail.as_view(),
          name="request-reset-email"),
-#     path('profile-update/', ProfileUpdateView.as_view(),
-#          name="profile-update"),
 ]
