+++ frontend/src/components/Breadcrumb.vue
@@ -1,0 +1,25 @@
+<template>
+    <!-- Breadcrumb Section Begin -->
+    <section class="breadcrumb-section set-bg" data-setbg="../assets/img/banner/banner-1.jpg">
+      <div class="container">
+        <div class="row">
+          <div class="col-lg-12 text-center">
+            <div class="breadcrumb__text">
+              <h2>Shop</h2>
+              <div class="breadcrumb__option">
+                <a href="/">Home</a>
+                <span>Shop</span>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+    <!-- Breadcrumb Section End -->
+</template>
+
+<script>
+export default {
+    name: "Breadcrumb"
+}
+</script>
