+++ frontend/src/views/Login.vue
@@ -83,7 +83,7 @@ export default {
 
           this.$store.commit("setToken", token);
 
-          axios.defaults.headers.common["Authorization"] = "Token " + token;
+          axios.defaults.headers.common["Authorization"] = `Token${token}`;
 
           localStorage.setItem("token", token);
