+++ frontend/src/views/Search.vue
@@ -1,63 +1,16 @@
 <template>
-  <div class="page-search">
-    <div class="columns is-multiline">
-      <div class="column is-12">
-        <h1 class="title">Search</h1>
-        <h2 class="is-size-5 has-text-grey">Search term: "{{ query }}"</h2>
-      </div>
-      <div class="row featured__filter" id="MixItUp556786" style="">
-        <FeaturedProduct
-          v-for="product in products"
-          :key="product.id"
-          :product="product"
-        />
+  <!-- Hero Section Begin -->
+  <section class="hero hero-normal">
+    <div class="container">
+      <div class="hero__search">
+        <div class="hero__search__form">
+          <form method="get" action="/search">
+            <input type="text" placeholder="What do you need?" name="query" />
+            <button type="submit" class="site-btn">SEARCH</button>
+          </form>
+        </div>
       </div>
     </div>
-  </div>
+  </section>
+  <!-- Hero Section End -->
 </template>
-
-<script>
-import axios from "axios";
-import FeaturedProduct from "@/components/FeaturedProduct.vue";
-
-export default {
-  name: "Search",
-  components: {
-    FeaturedProduct,
-  },
-  data() {
-    return {
-      products: [],
-      query: "",
-    };
-  },
-  mounted() {
-    document.title = "Search";
-
-    let uri = window.location.search.substring(1);
-    let params = new URLSearchParams(uri);
-
-    if (params.get("query")) {
-      this.query = params.get("query");
-
-      this.performSearch();
-    }
-  },
-  methods: {
-    async performSearch() {
-      this.$store.commit("setIsLoading", true);
-
-      await axios
-        .post("/api/v1/products/search/", { query: this.query })
-        .then((response) => {
-          this.products = response.data;
-        })
-        .catch((error) => {
-          console.log(error);
-        });
-
-      this.$store.commit("setIsLoading", false);
-    },
-  },
-};
-</script>
